ya funcionan las redirecciones solo falta una arreglar en nav y el header

+++ lib/feature/publication/data/publication_api_data_source.dart
@@ -2,6 +2,7 @@ import 'dart:convert';
 import 'package:http/http.dart' as http;
 import 'dart:async';
 import 'package:actividad1c2/feature/publication/data/publication_model.dart';
+import 'package:shared_preferences/shared_preferences.dart';
 
 import '../domain/publication.dart';
 
@@ -19,10 +20,15 @@ class PublicationApiDataSourceImp implements PublicationApiDataSource {
   @override
   Future<void> createPublication(
       String idUser, String description, dynamic urlFile) async {
+         String? token; // Variable para almacenar el token recuperado
     try {
+      SharedPreferences prefs = await SharedPreferences.getInstance();
+      token = prefs.getString('token');
+       if (token == null) {
+        throw Exception('No token found! Unable to send the request.');
+      }
       var headers = {
-        'Authorization':
-            'Bearer eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJ1dWlkIjoiZDZlNGJlNTYtZDk4MC00YTE5LWE5NjAtZDNhY2ZjNzdjMThmIiwiZW1haWwiOiJqdWFuMTJAZXhhbXBsZS5jb20iLCJpYXQiOjE2OTgwODc1MjQsImV4cCI6MTY5ODM1MzkyNH0.2YW_ueR2faKIeHJfYshh13JNrL241cfYefEKeAYZXDE', // Coloca aquí tu token actual
+        'Authorization': 'Bearer $token'
       };
 
       var request = http.MultipartRequest(
@@ -46,6 +52,7 @@ class PublicationApiDataSourceImp implements PublicationApiDataSource {
 
       if (response.statusCode == 200) {
         print(await response.stream.bytesToString());
+        
       } else {
         print(response.reasonPhrase);
       }
@@ -56,34 +63,42 @@ class PublicationApiDataSourceImp implements PublicationApiDataSource {
   }
 
   @override
-  Future<void> deletePublication(String uuid) async {
+Future<void> deletePublication(String uuid) async {
+    // Recupera las SharedPreferences
+    SharedPreferences prefs = await SharedPreferences.getInstance();
+    // Intenta obtener el token del almacenamiento local
+    String? token = prefs.getString('token');
+
+    // Si el token es null, debemos lanzar un error
+    if (token == null) {
+      throw Exception('Token is null. Please re-login or obtain a new token.');
+    }
+
+    // Prepara los encabezados de la solicitud con el token
     var headers = {
-      'Authorization':
-          'Bearer eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJ1dWlkIjoiNDhlZjMzZTAtNWU1ZS00NDcxLWI3NjMtYjMyZWQ0Y2YxYjA0IiwiZW1haWwiOiJqdWFuMUBleGFtcGxlLmNvbSIsImlhdCI6MTY5ODIyMDU1OCwiZXhwIjoxNjk4NDg2OTU4fQ.O0FvXI1wMJMF8hTgAVvJLazXjIpWenVaHQo7AIopd4s'
+      'Authorization': 'Bearer $token', // usa el token real desde SharedPreferences
     };
 
-    var url = Uri.parse(
-        'https://actual-servant-production.up.railway.app/api/v1/public/delete/$uuid'); // Asegúrate de que tu endpoint puede manejar la variable uuid.
+    var url = Uri.parse('https://actual-servant-production.up.railway.app/api/v1/public/delete/$uuid'); 
 
-    // Usando un try-catch para manejar posibles errores durante la solicitud.
     try {
       var response = await http.delete(url, headers: headers);
 
       if (response.statusCode == 200) {
         print('Publicación eliminada con éxito. Respuesta: ${response.body}');
       } else {
-        print(
-            'Error al eliminar la publicación. Estado: ${response.statusCode}. Razón: ${response.reasonPhrase}');
+        print('Error al eliminar la publicación. Estado: ${response.statusCode}. Razón: ${response.reasonPhrase}');
+        // Aquí puedes lanzar una excepción o manejar este estado de error como prefieras
       }
     } catch (e) {
-      // Esto atrapará cualquier otra excepción que pueda ocurrir.
-      // Puedes manejarlo de manera diferente, dependiendo de tus necesidades.
       print('Ocurrió un error al enviar la solicitud: $e');
+      // Aquí también puedes lanzar la excepción para manejarla en una capa superior
+      throw Exception('Error deleting publication: $e');
     }
   }
-
   @override
-  Future<Publication> getPublicationByUser(String name) async {//no se usa 
+  Future<Publication> getPublicationByUser(String name) async {
+    //no se usa
     // Construir el URI usando el nombre proporcionado
     var uri = Uri.parse('http://localhost:3001/api/v1/public/user/$name');
 
@@ -117,7 +132,8 @@ class PublicationApiDataSourceImp implements PublicationApiDataSource {
   }
 
   @override
-  Future<Publication> getPublicationByUuid(String uuid) async {//no se usa 
+  Future<Publication> getPublicationByUuid(String uuid) async {
+    //no se usa
     // Construir el URI usando el UUID proporcionado
     var uri = Uri.parse('http://localhost:3001/api/v1/public/$uuid');
 
@@ -152,68 +168,91 @@ class PublicationApiDataSourceImp implements PublicationApiDataSource {
 
   @override
   Future<List<Publication>> listPublication() async {
+        // Inicializando el SharedPreferences
+    SharedPreferences prefs = await SharedPreferences.getInstance();
+    String? token = prefs.getString('token');
+
+    // Si el token es null, debemos lanzar un error
+    if (token == null) {
+      throw Exception('Token is null. Please re-login or obtain a new token.');
+    }
+
     var headers = {
-      'Authorization':
-          'Bearer eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJ1dWlkIjoiZDZlNGJlNTYtZDk4MC00YTE5LWE5NjAtZDNhY2ZjNzdjMThmIiwiZW1haWwiOiJqdWFuMTJAZXhhbXBsZS5jb20iLCJpYXQiOjE2OTgwODc1MjQsImV4cCI6MTY5ODM1MzkyNH0.2YW_ueR2faKIeHJfYshh13JNrL241cfYefEKeAYZXDE'
+      'Authorization': 'Bearer $token', // utiliza el token real desde SharedPreferences
     };
-    print("entre a listar");
-    var response = await http.get(
-      Uri.parse(
-          'https://actual-servant-production.up.railway.app/api/v1/public/all'),
-      headers: headers,
-    );
 
-    if (response.statusCode == 200) {
-      // Decodifica la respuesta JSON
-      Map<String, dynamic> jsonData = jsonDecode(response.body);
+    try {
+      print("entre a listar");
+      var response = await http.get(
+        Uri.parse('https://actual-servant-production.up.railway.app/api/v1/public/all'),
+        headers: headers,
+      );
 
-      // Accede al campo 'data' y luego al campo 'list' que contiene las publicaciones
-      List<dynamic> publicationsList = jsonData['data']['list'];
+      if (response.statusCode == 200) {
+        // Decodifica la respuesta JSON
+        Map<String, dynamic> jsonData = jsonDecode(response.body);
 
-      // Convierte cada elemento en la lista a un objeto 'Publication'.
-      List<Publication> publications = publicationsList
-          .map((dynamic item) => PublicationModel.fromJson(item))
-          .toList();
+        // Accede al campo 'data' y luego al campo 'list' que contiene las publicaciones
+        List<dynamic> publicationsList = jsonData['data']['list'];
 
-      return publications;
-    } else {
-      throw Exception('Failed to load publications from API');
+        // Convierte cada elemento en la lista a un objeto 'Publication'.
+        List<Publication> publications = publicationsList
+            .map((dynamic item) => PublicationModel.fromJson(item))
+            .toList();
+
+        return publications;
+      } else {
+        // Si el servidor responde con un código de error, lanzamos una excepción
+        throw Exception('Failed to load publications from API with status code ${response.statusCode}');
+      }
+    } catch (error) {
+      // Esto atrapará cualquier error de la solicitud y lo lanzará para ser manejado por el código que llama al método
+      print(error.toString());
+      throw Exception('Failed to load publications: $error');
     }
   }
 
   @override
   Future<void> updateDescription(String uuid, String description) async {
-    var headers = {
-      'Content-Type': 'application/json',
-      'Authorization': 'Bearer eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJ1dWlkIjoiZDZlNGJlNTYtZDk4MC00YTE5LWE5NjAtZDNhY2ZjNzdjMThmIiwiZW1haWwiOiJqdWFuMTJAZXhhbXBsZS5jb20iLCJpYXQiOjE2OTgwODc1MjQsImV4cCI6MTY5ODM1MzkyNH0.2YW_ueR2faKIeHJfYshh13JNrL241cfYefEKeAYZXDE', // reemplaza con tu token real
-    };
+      SharedPreferences prefs = await SharedPreferences.getInstance();
+  String? token = prefs.getString('token');
 
-    var request = http.Request(
-        'PUT', Uri.parse('https://actual-servant-production.up.railway.app/api/v1/public/description'));
-
-    // Asegúrate de que el cuerpo de la solicitud contenga los campos que espera tu API.
-    var body = json.encode({
-      "uuid": uuid,
-      "description": description,
-    });
-
-    request.body = body;
-    request.headers.addAll(headers);
-
-    try {
-      http.StreamedResponse response = await request.send();
-
-      if (response.statusCode == 200) {
-        print(await response.stream.bytesToString());
-      } else {
-        // Aquí puedes manejar errores de diferentes maneras, no solo imprimir
-        print(
-          'Error al actualizar la descripción: ${response.reasonPhrase}',
-        );
-      }
-    } catch (e) {
-      // Manejo de errores de red, etc.
-      print('Error al realizar la solicitud: $e');
-    }
+  // Si el token no existe, es una buena práctica lanzar un error o manejar esta situación de alguna manera.
+  if (token == null) {
+    throw Exception("No se encontró el token. Es necesario iniciar sesión de nuevo.");
   }
-}
+
+  var headers = {
+    'Content-Type': 'application/json',
+    'Authorization': 'Bearer $token', // Aquí se usa el token recuperado de SharedPreferences
+  };
+
+  var request = http.Request(
+    'PUT',
+    Uri.parse('https://actual-servant-production.up.railway.app/api/v1/public/description')
+  );
+
+  var body = json.encode({
+    "uuid": uuid,
+    "description": description,
+  });
+
+  request.body = body;
+  request.headers.addAll(headers);
+
+  try {
+    http.StreamedResponse response = await request.send();
+
+    if (response.statusCode == 200) {
+      print(await response.stream.bytesToString());
+    } else {
+      // En caso de un código de estado diferente a 200, se debe manejar el error.
+      print('Error al actualizar la descripción. Estado: ${response.statusCode}.');
+      throw Exception('Error al actualizar la descripción: ${response.reasonPhrase}');
+    }
+  } catch (e) {
+    // Si se produce una excepción durante la solicitud, se atrapa aquí.
+    print('Error al realizar la solicitud: $e');
+    throw Exception('Error al realizar la solicitud: $e');
+  }
+}}
